Move logic around to different scripts
