+++ index.html
@@ -31,9 +31,10 @@
     <link rel="preload" href="/fonts/JosefinSans.woff2" as="font" type="font/woff2" crossorigin>
     <link rel="preload" href="/fonts/OxygenMono.woff2" as="font" type="font/woff2" crossorigin>
 
-    <script type="module" src="/src/index.ts"></script>
     <script type="module" src="/src/theme.ts"></script>
     <script type="module" src="/src/logo-web-component.ts"></script>
+    <script type="module" src="/src/projects.ts"></script>
+    <script type="module" src="/src/misc.ts"></script>
 </head>
 
 <body>
