+++ src/projects.ts
@@ -1,6 +1,7 @@
 import domPurify from "dompurify";
 import * as githubColorsJson from "./github-colors.json";
 import { Logo } from "./logo-web-component";
+import { fetchData, fetchJson } from "./utils";
 
 type Site = "Cults 3D" | "Github" | "Board Game Geek";
 type Language = { name: string; color: string };
@@ -287,3 +288,49 @@ export const projectIntoTemplate = (
 
     return templateClone;
 };
+const loadProjects = async () => {
+    // Create project-gallery loader
+    const loader = document.createElement("span");
+    loader.classList.add("loader");
+    document.getElementById("projects")?.append(loader);
+
+    // Load items into gallery
+    const gallery = document.getElementById("project-gallery");
+    const template = document.getElementById("project-template") as HTMLTemplateElement | undefined;
+
+    if (gallery && template) {
+        const githubPage = await fetchJson<GithubRepo[]>("/proxy/api/github");
+        const githubProjects = githubRepoToProject(githubPage).map((p) =>
+            projectIntoTemplate(p, template),
+        );
+        gallery.append(...githubProjects);
+
+        const bggPage = await fetchData("/proxy/boardgamegeek");
+        const bggRawProjects = scrapeBgg(bggPage);
+
+        // Get higher resolution image from bgg xmlapi
+        for (const project of bggRawProjects) {
+            const id = project.url
+                ?.toString()
+                ?.split("/")
+                .find((v) => v.match(/\d+/g));
+            const gameXml = await fetchData(`/proxy/xmlapi/boardgamegeek/${id}`, "text/xml");
+            upgradeBggData(project, gameXml);
+        }
+
+        const bggProjects = bggRawProjects.map((p) => projectIntoTemplate(p, template));
+
+        gallery.append(...bggProjects);
+
+        const cults3dPage = await fetchData("/proxy/cults3d");
+        const cults3dProjects = scrapeCults3d(cults3dPage).map((p) =>
+            projectIntoTemplate(p, template),
+        );
+        gallery.append(...cults3dProjects);
+    }
+
+    // remove project-gallery loader
+    loader.remove();
+};
+
+loadProjects();
